msauth: implement workaround for MSA-PT silent auth
If we have an MSA pass-through (MSA-PT) resource/scope and an existing
user is an MSA, then we need to do some 'magic' to get a valid access
token for the resource.

MSA accounts have a pseudo 'tenant ID' that an IAccount object has as
it's home tenant ID.  The MSA-PT resource however requires that they
have a token issued from a 'real' AAD tenant.  There are special AAD
tenants for each Azure cloud instance where every MSA is a guest, and we
must use this tenant instead when doing a silent token request.

Yes.. I know...