fs: fix IsSamePath bug
Fix the same bug in all implementations of the IFileSystem::IsSamePath()
method. We have been accidentially only comparing the file/basenames of
paths for equality, rather than the entire path!

Update all implementations to compare the entire file path. Also add
unit tests to catch this problem in the future.