msauth: pull up PCA build to calling methods
Pull the building of the MSAL public client application up one level to
the calling method (GetTokenAsync) in prep for adding additional methods
that need to create a PCA in a slightly different way.

As part of this change, also simplify initalisation of the broker and
construction of the PCA with the broker.