WIP dotnet: migrate all projs to net6.0; drop netfx
Drop .NET Framework as a target framework on Windows and migrate all
projects to only .NET 6.0 (also dropping .NET Standard).

At the same time, drop all of the #if NETFRAMEWORK conditional
compilation and inline/replace PlatformUtils with the OperatingSystem
class that is now available everywhere.

In dropping .NET Framework on Windows, we also adopt the new
"MSALRuntime" WAM broker for MSAL/Microsoft Authentication, to avoid the
mess of dual targeting netX-windowsY.