wsl: add ability to detect WSL distro and version
Add methods to the WslUtils class to enable detection of a WSL
distribution, and also determine which WSL version is being used.

Version 1 uses the Windows NT kernel and runs the distribution in the
same user-mode space as Windows programs.

Version 2 uses a light-weight VM to host a real Linux kernel, and runs
the distribution also inside the VM; interop with Windows is achieved
using other mechanisms.